+++ include/stxxl/bits/algo/sort_base.h
@@ -40,7 +40,7 @@ __STXXL_BEGIN_NAMESPACE
 // Optimal merging: merge r = pow(nruns,1/ceil(log(nruns)/log(m))) runs at once
 inline unsigned_type optimal_merge_factor(unsigned_type num_runs, unsigned_type max_concurrent_runs)
 {
-    return unsigned_type(ceil(pow(num_runs, 1. / ceil(log(double(num_runs)) / log(double(max_concurrent_runs))))));
+    return unsigned_type(ceil(pow(double(num_runs), 1. / ceil(log(double(num_runs)) / log(double(max_concurrent_runs))))));
 }
 
 __STXXL_END_NAMESPACE
